refactor(BookInfoPopup): move into a new module
