+++ index.html
@@ -6,7 +6,11 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
     <title>Document</title>
+
+    <!-- styles -->
     <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="src/BookInfoPopup/BookInfoPopup.css" />
+    <link rel="stylesheet" href="src/ExitButton/ExitButton.css" />
 
     <!-- fonts -->
     <link rel="preconnect" href="https://fonts.googleapis.com" />
