+++ style.css
@@ -55,35 +55,6 @@ button:active {
   box-shadow: var(--small-shadow-inset);
 }
 
-.exit {
-  position: relative;
-
-  height: 20px;
-  width: 20px;
-
-  cursor: pointer;
-}
-
-.exit .line {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  translate: -50% -50%;
-
-  height: 2px;
-  width: 20px;
-  background-color: var(--bg-color);
-  rotate: 45deg;
-}
-
-.exit .line:nth-child(2) {
-  rotate: -45deg;
-}
-
-.exit:hover .line {
-  background-color: white;
-}
-
 body {
   background-color: var(--bg-color);
   font-family: "Poppins", sans-serif;
